# Multiply64 optimized with Math.BigMul