@Scooletz 2021-02-11
Ping @tkstanczak . This might be best `let's just use .NET 5 method instead` that I had in a while 😅